ProjectService_DbContext (#2)
* Initial Create: ProjectService

Added a new empty Web Application project "ProjectService" to the "DigitalOffice" solution.

* DbContext and Entity definition

Created a new folder "Database" where i defined DbContext, creating a new class "ProjectServiceDbContext". In there I defined the table it'll be using -- Projects.

In the "Database" folder created a new one, called "Entities". There I created a class that defines the table "Projects". Called it "Project.cs".

In "Project.cs" I used two attributes: [Key] and [NotMapped]
Key is for the primary key definition
NotMapped is for excluding the property from the table.
I got confused with the implementation of these fields because we're working with the separated databases.

* DbContext registration

Added a "ConnectionStrings" section in the "appsettings.json" file.
It contains a first entry called "SQLConnectionString" that is used for connecting to the SQL Server.

In "Startup.cs" I added "Configuration" field so that I can access the configuration file.
DbContext is registered via "services.AddDbContext".

* First DbContext Migration

Using PM console, typed:
1. Add-Migration InitialCreate
2. Update-Database -Migration InitialCreate.

It created a new folder "Migrations", captured all the definitions and then updated the database. In this case, database did not exist, so EntityFrameworkCore created it using the provided in config file "SQLConnectionString".

* Updated ports

Ports are changed from 5000 to different ones so that in the future we'll be able to run multiple projects simultaneously

* New folder "Services", new localhost ports, added .UseHttpRedirection

Moved ProjectService files to a new folder "Services"

----
Also updated ports and added .UseHttpRedirection() to Startup.cs

* Updated Entities and added 2 new ones

Created 2 new models:
ProjectWorkerUser.cs and ProjectManagerUser.cs that serve as a List<Guid> for the Projects table.
These new tables contain pairs of ProjectId and WorkerUserId/ManagerUserId

Migrations are applied and the database is updated. Everything's in order.

* Automated Migrations

Migrations are applied to the addressed DbContext in a new private static method that gets a scope of the dbContext and updates its database.

* Added ProjectFile Entity

Created a new entity called "ProjectFile" that serves as a link between Projects and its Files.
That table contains pairs of Ids -- ProjectId + FileId.

Migration created, applied to the database.

Co-Authored-By: Evgeniy <38541098+djino97@users.noreply.github.com>

* Obsolete property --> IsActive property

Renamed the "disabling" property

Co-Authored-By: Evgeniy <38541098+djino97@users.noreply.github.com>

* Update DigitalOffice.sln

Added a solution folder "Services", moved ProjectService there

* Deleted "Services" Folder

Moved ProjectService files from "Services" folder to "Services" SOLUTION folder.

On the disk it'll seem as they are in different folders, but in the solution there'll be a "Services" folder

* Removed blank spaces, moved ProjectService to "Services" folder (FINAL DECISION)

Co-authored-by: Evgeniy <38541098+djino97@users.noreply.github.com>