Entity Update
We added entities for many-to-many relationships:
*Relationship between User and Position.
Position or User shouldn't include information about Start and End Time for position itself, because we will store extra data there(too much not very different entries). So we moved these fields(StartTime, EndTime)  to relationship entity.
*Relationship between User and File(Certificate).
*Relationship between User and Achievement.
Achievement or User shouldn't include information about Time of getting an achievement itself, so it must be in their relationship entity, just like in UserPosition model. And we moved this field(Time) to relationship model.